+++ axelrod/fingerprint.py
@@ -1,20 +1,21 @@
 import os
 from collections import namedtuple
 from tempfile import mkstemp
-from typing import Any, List, Union
+from typing import Any, List, Optional, Union
 
-import axelrod as axl
 import dask.dataframe as dd
 import matplotlib.pyplot as plt
 import numpy as np
 import tqdm
+from mpl_toolkits.axes_grid1 import make_axes_locatable
+
+import axelrod as axl
 from axelrod import Player
 from axelrod.interaction_utils import (
     compute_final_score_per_turn,
     read_interactions_from_file,
 )
 from axelrod.strategy_transformers import DualTransformer, JossAnnTransformer
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 Point = namedtuple("Point", "x y")
 
@@ -42,7 +43,7 @@ def _create_points(step: float, progress_bar: bool = True) -> List[Point]:
     num = int((1 / step) // 1) + 1
 
     if progress_bar:
-        p_bar = tqdm.tqdm(total=num ** 2, desc="Generating points")
+        p_bar = tqdm.tqdm(total=num**2, desc="Generating points")
 
     points = []
     for x in np.linspace(0, 1, num):
@@ -279,10 +280,10 @@ class AshlockFingerprint(object):
         turns: int = 50,
         repetitions: int = 10,
         step: float = 0.01,
-        processes: int = None,
-        filename: str = None,
+        processes: Optional[int] = None,
+        filename: Optional[str] = None,
         progress_bar: bool = True,
-        seed: int = None,
+        seed: Optional[int] = None,
     ) -> dict:
         """Build and play the spatial tournament.
 
@@ -357,7 +358,7 @@ class AshlockFingerprint(object):
         self,
         cmap: str = "seismic",
         interpolation: str = "none",
-        title: str = None,
+        title: Optional[str] = None,
         colorbar: bool = True,
         labels: bool = True,
     ) -> plt.Figure:
@@ -436,11 +437,11 @@ class TransitiveFingerprint(object):
         self,
         turns: int = 50,
         repetitions: int = 1000,
-        noise: float = None,
-        processes: int = None,
-        filename: str = None,
+        noise: Optional[float] = None,
+        processes: Optional[int] = None,
+        filename: Optional[str] = None,
         progress_bar: bool = True,
-        seed: int = None,
+        seed: Optional[int] = None,
     ) -> np.ndarray:
         """Creates a spatial tournament to run the necessary matches to obtain
         fingerprint data.
@@ -555,7 +556,7 @@ class TransitiveFingerprint(object):
         self,
         cmap: str = "viridis",
         interpolation: str = "none",
-        title: str = None,
+        title: Optional[str] = None,
         colorbar: bool = True,
         labels: bool = True,
         display_names: bool = False,