+++ axelrod/match.py
@@ -56,41 +56,39 @@ class Match(object):
             Random seed for reproducibility
         """
 
-        defaults = {
-            (True, True): (DEFAULT_TURNS, 0),
-            (True, False): (float("inf"), prob_end),
-            (False, True): (turns, 0),
-            (False, False): (turns, prob_end),
-        }
-        self.turns, self.prob_end = defaults[(turns is None, prob_end is None)]
+        self.turns, self.prob_end = turns, prob_end
+        if prob_end is None:
+            self.prob_end = 0
+        if turns is None:
+            self.turns = float("inf")
+        if turns is None and prob_end is None:
+            self.turns = DEFAULT_TURNS
 
         self.result = []
         self.noise = noise
 
-        self.set_seed(seed)
-
+        self.game = game
         if game is None:
             self.game = Game()
-        else:
-            self.game = game
 
+        self._cache = deterministic_cache
         if deterministic_cache is None:
             self._cache = DeterministicCache()
-        else:
-            self._cache = deterministic_cache
 
+        self.match_attributes = match_attributes
         if match_attributes is None:
+            # known_turns = inf if both prob_end and turns are None, else turns
             known_turns = self.turns if prob_end is None else float("inf")
             self.match_attributes = {
                 "length": known_turns,
                 "game": self.game,
                 "noise": self.noise,
             }
-        else:
-            self.match_attributes = match_attributes
 
         self.players = list(players)
         self.reset = reset
+
+        self.set_seed(seed)
 
     def set_seed(self, seed):
         """Sets a random seed for the Match, for reproducibility. Initializes
@@ -130,7 +128,8 @@ class Match(object):
         A boolean to show whether the deterministic cache should be updated.
         """
         return (
-            not self.noise
+            not self._stochastic
+            and not self.noise
             and self._cache.mutable
             and not (any(Classifiers["stochastic"](p) for p in self.players))
         )
@@ -182,28 +181,29 @@ class Match(object):
             turns = self.turns
         cache_key = (self.players[0], self.players[1])
 
-        if self._stochastic or not self._cached_enough_turns(cache_key, turns):
-            for p in self.players:
-                if self.reset:
-                    p.reset()
-                p.set_match_attributes(**self.match_attributes)
-                # Generate a random seed for the player, if stochastic
-                if Classifiers["stochastic"](p):
-                    p.set_seed(self._random.random_seed_int())
-            result = []
-            for _ in range(turns):
-                plays = self.simultaneous_play(
-                    self.players[0], self.players[1], self.noise
-                )
-                result.append(plays)
+        if self._cached_enough_turns(cache_key, turns):
+            self.result = self._cache[cache_key][:turns]
+            return self.result
 
-            if self._cache_update_required:
-                self._cache[cache_key] = result
-        else:
-            result = self._cache[cache_key][:turns]
+        for p in self.players:
+            if self.reset:
+                p.reset()
+            p.set_match_attributes(**self.match_attributes)
+            # Generate a random seed for the player, if stochastic
+            if Classifiers["stochastic"](p):
+                p.set_seed(self._random.random_seed_int())
+        result = []
+        for _ in range(turns):
+            plays = self.simultaneous_play(
+                self.players[0], self.players[1], self.noise
+            )
+            result.append(plays)
+
+        if self._cache_update_required:
+            self._cache[cache_key] = result
 
         self.result = result
-        return result
+        return self.result
 
     def scores(self):
         """Returns the scores of the previous Match plays."""