+++ axelrod/tests/unit/test_tournament.py
@@ -1,4 +1,5 @@
 """Tests for the main tournament class."""
+
 import io
 import logging
 import os
@@ -9,9 +10,13 @@ import warnings
 from multiprocessing import Queue, cpu_count
 from unittest.mock import MagicMock, patch
 
-import axelrod as axl
 import numpy as np
 import pandas as pd
+from hypothesis import example, given, settings
+from hypothesis.strategies import floats, integers
+from tqdm import tqdm
+
+import axelrod as axl
 from axelrod.load_data_ import axl_filename
 from axelrod.tests.property import (
     prob_end_tournaments,
@@ -19,10 +24,7 @@ from axelrod.tests.property import (
     strategy_lists,
     tournaments,
 )
-from axelrod.tournament import _close_objects
-from hypothesis import example, given, settings
-from hypothesis.strategies import floats, integers
-from tqdm import tqdm
+from axelrod.tournament import MatchChunk, _close_objects
 
 C, D = axl.Action.C, axl.Action.D
 
@@ -661,7 +663,9 @@ class TestTournament(unittest.TestCase):
                 for player2_index in range(player1_index, len(self.players)):
                     index_pair = (player1_index, player2_index)
                     match_params = {"turns": turns, "game": self.game}
-                    yield (index_pair, match_params, self.test_repetitions, 0)
+                    yield MatchChunk(
+                        index_pair, match_params, self.test_repetitions, 0
+                    )
 
         chunk_generator = make_chunk_generator()
         interactions = {}