+++ docs/tutorials/running_axelrods_first_tournament/_static/running_axelrods_first_tournament/main.py
@@ -2,8 +2,9 @@
 Script to obtain plots for the running axelrod tournament tutorial.
 """
 
-import axelrod as axl
 import matplotlib.pyplot as plt
+
+import axelrod as axl
 
 first_tournament_participants_ordered_by_reported_rank = [
     s() for s in axl.axelrod_first_strategies
@@ -11,11 +12,11 @@ first_tournament_participants_ordered_by_reported_rank = [
 number_of_strategies = len(
     first_tournament_participants_ordered_by_reported_rank
 )
-axl.seed(0)
 tournament = axl.Tournament(
     players=first_tournament_participants_ordered_by_reported_rank,
     turns=200,
     repetitions=5,
+    seed=0,
 )
 results = tournament.play()